The monitor does not halt simulation any more.

+++ lab1/TrafficSystem_testbench/constants.cc
@@ -1,5 +1,5 @@
 
 #include "constants.h"
 
-const int Constants::gen_time = 30;
-const int Constants::add_time = 10;
+const double Constants::gen_time = 30;
+const double Constants::add_time = 10;
